refactor: updated GitHub Scraper w/ our upstream changes
